Added superclass for handling oauth in clients
